Disable ad display and update package references
The ad display function `showAd()` in `ForgeInstaller.cs` has been commented out to disable showing of ads. In addition, a package reference for "CmlLib.Core" in `CmlLib.Core.Installer.Forge.csproj` was removed and a new project reference to "CmlLib.Core.csproj" was added. This change will ensure the use of the most recent and relevant references.