Add method to get Forge versions for specific Minecraft version (#5)
* Add method to get Forge versions for specific Minecraft version

A new public method, `GetForgeVersions`, has been added to the `ForgeInstaller.cs` class. This method returns the available Forge versions for a specific Minecraft version, improving the process of version management and installation.

* Disable ad display and update package references

The ad display function `showAd()` in `ForgeInstaller.cs` has been commented out to disable showing of ads. In addition, a package reference for "CmlLib.Core" in `CmlLib.Core.Installer.Forge.csproj` was removed and a new project reference to "CmlLib.Core.csproj" was added. This change will ensure the use of the most recent and relevant references.

* Revert "Disable ad display and update package references"

This reverts commit 9040df87a6d1f25b37bf4718fcf77634110adc66.